requirements.txt
asgiref==3.3.2
Django==3.2.15
django-cors-headers==4.3.0
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-rest-passwordreset==1.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==21.2.0
jsonfield==3.1.0
mysqlclient==2.2.4
packaging==23.2
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2019.3
sqlparse==0.4.4
swapper==1.3.0
